+++ R/dimensions.R
@@ -674,8 +674,13 @@ print.dimensions = function(x, ...) {
 identical_dimensions = function(lst, ignore_resolution = FALSE, tolerance = 0) {
 	if (length(lst) > 1) {
 		d1 = attr(lst[[1]], "dimensions")
+		crs1 = st_crs(d1)
+		st_crs(d1) = NA_crs_
 		for (i in 2:length(lst)) {
 			di = attr(lst[[i]], "dimensions")
+			if (st_crs(di) != crs1) # check semantical equivalence; https://github.com/r-spatial/stars/issues/703
+				return(FALSE)
+			st_crs(di) = NA_crs_
 			if (ignore_resolution) {
 				for (j in seq_along(d1))
 					d1[[j]]$delta = d1[[j]]$to = NA_real_